get vhd parsing to work

+++ src/cpu/cpu.h
@@ -55,7 +55,7 @@ int32_t cpu_exec86(int32_t cycle_target);
 // get the current tick count of this slice
 uint64_t cpu_slice_ticks(void);
 
-typedef void (*cpu_intcall_t)(const uint8_t int_num);
+typedef void (*cpu_intcall_t)(const uint16_t int_num);
 void cpu_set_intcall_handler(cpu_intcall_t handler);
 
 // cpu instruction pointer
